Move cli mod to be a child of the api mod
A cli is a form of api, so it may make sense to move the
cli to the api mod. This is similar to what laravel does, iirc.

To avoid introducing breaking changes, introduce re-exports of the moved
structs/enums/etc from their previous locations, and mark the old mod as
deprecated. The old mod will be removed in a future api-breaking version
(`0.3.0+`)

Closes https://github.com/roadster-rs/roadster/issues/169
