+++ src/api/cli/roadster/list_routes.rs
@@ -1,0 +1,5 @@
+use clap::Parser;
+use serde_derive::Serialize;
+
+#[derive(Debug, Parser, Serialize)]
+pub struct ListRoutesArgs {}
