+++ src/api/mod.rs
@@ -1,2 +1,4 @@
+#[cfg(feature = "cli")]
+pub mod cli;
 #[cfg(feature = "http")]
 pub mod http;
